==> notebooks/01_build_mlb_roles.ipynb <==
import pandas as pd
print("Pandas version:", pd.__version__)

fielding = pd.read_csv("../data/raw/mlb/fielding.csv")
fielding.head()

fielding.columns

# 1. Convert Outs to Innings
fielding['Innings'] = fielding['InnOuts'] / 3

# 2. Aggregate to player-season-position
mlb_def_roles = (
    fielding
    .groupby(['playerID', 'yearID', 'POS'], as_index=False)
    .agg(time_in_role = ('Innings', 'sum'))
    .rename(columns={
        'playerID': 'player_id',
        'yearID': 'season',
        'POS': 'position'
    })
)

# 3. Add "League" column
mlb_def_roles['league'] = 'MLB'

mlb_def_roles.head()

mlb_def_roles["total_time"] = (
    mlb_def_roles
    .groupby(["player_id", "season"])["time_in_role"]
    .transform("sum")
)

mlb_def_roles["role_share"] = mlb_def_roles["time_in_role"] / mlb_def_roles["total_time"]

(mlb_def_roles
 .groupby(["player_id", "season"])["role_share"]
 .sum()
 .describe()
)

batting = pd.read_csv("../data/raw/mlb/batting.csv")
pitching = pd.read_csv("../data/raw/mlb/pitching.csv")

# -----------------------
# HITTER ROLES (PA-weighted)
# -----------------------
# Making sure missing columns don't break things
for c in ["BB", "HBP", "SH", "SF"]:
    if c not in batting.columns:
        batting[c] = 0

bat_agg = (
    batting
    .groupby(["playerID", "yearID"], as_index=False)
    .agg(AB=("AB", "sum"),
         BB=("BB", "sum"),
         HBP=("HBP", "sum"),
         SH=("SH", "sum"),
         SF=("SF", "sum"))
)

bat_agg["PA"] = bat_agg["AB"] + bat_agg["BB"] + bat_agg["HBP"] + bat_agg["SH"] + bat_agg["SF"]

hitter_roles = (
    bat_agg.loc[bat_agg["PA"] > 0, ["playerID", "yearID", "PA"]]
    .rename(columns={"playerID": "player_id", "yearID": "season", "PA": "time_in_role"})
)
hitter_roles["position"] = "HITTER"
hitter_roles["league"] = "MLB"

hitter_roles = hitter_roles[["player_id", "season", "position", "time_in_role", "league"]]

# -----------------------
# PITCHER ROLES (IP-weighted)
# -----------------------
# Lahman Pitching uses IPouts (outs recorded). Convert to innings.
pit_agg = (
    pitching
    .groupby(["playerID", "yearID"], as_index=False)
    .agg(IPouts=("IPouts", "sum"))
)

pitcher_roles = pit_agg.copy()
pitcher_roles["time_in_role"] = pitcher_roles["IPouts"] / 3.0
pitcher_roles = pitcher_roles.loc[pitcher_roles["time_in_role"] > 0, ["playerID", "yearID", "time_in_role"]]
pitcher_roles = pitcher_roles.rename(columns={"playerID": "player_id", "yearID": "season"})
pitcher_roles["position"] = "P"
pitcher_roles["league"] = "MLB"

pitcher_roles = pitcher_roles[["player_id", "season", "position", "time_in_role", "league"]]

# -----------------------
# COMBINE
# -----------------------
mlb_roles = pd.concat([mlb_def_roles, hitter_roles, pitcher_roles], ignore_index=True)

mlb_roles.head(), hitter_roles.head(), pitcher_roles.head()

import numpy as np

def entropy(shares):
    shares = shares[shares > 0]
    return -np.sum(shares * np.log(shares))

utility_entropy = (
    mlb_roles
    .groupby(["player_id", "season"])
    .agg(
        utility_entropy=("role_share", entropy),
        n_roles=("position", "nunique")
    )
    .reset_index()
)

utility_entropy["max_entropy"] = np.log(utility_entropy["n_roles"])
utility_entropy["utility_score"] = (
    utility_entropy["utility_entropy"] / utility_entropy["max_entropy"]
)

# handle n_roles = 1
utility_entropy.loc[utility_entropy["n_roles"] == 1, "utility_score"] = 0.0


utility_entropy.head()

utility_entropy["utility_score"] = utility_entropy["utility_score"].clip(lower=0)

mlb_roles = mlb_roles.merge(
    utility_entropy[["player_id", "season", "utility_score"]],
    on=["player_id", "season"],
    how="left"
)

mlb_roles.to_csv("mlb_roles_with_utility.csv", index=False)
utility_entropy.to_csv("season_utility_scores.csv", index=False)
